LPS-167636 Move Table to common
